WIP add nested field support
