+++ packages/data-types/src/types/versions/base-type.ts
@@ -3,16 +3,25 @@ import * as ts from '@terascope/utils';
 import { TypeESMapping, GraphQLType, FieldTypeConfig } from '../../interfaces';
 import { formatGQLComment } from '../../graphql-helper';
 
+export type NestedTypes = { [field: string]: BaseType };
+
+// eslint-disable-next-line @typescript-eslint/interface-name-prefix
+export interface IBaseType {
+    new(field: string, config: FieldTypeConfig, nestedTypes?: NestedTypes): BaseType;
+}
+
 export default abstract class BaseType {
     readonly field: string;
+    readonly nestedTypes?: NestedTypes;
     readonly config: FieldTypeConfig;
 
-    constructor(field: string, config: FieldTypeConfig) {
+    constructor(field: string, config: FieldTypeConfig, nestedTypes?: NestedTypes) {
         if (!field || !ts.isString(field)) {
             throw new ts.TSError('A field must be provided and must be of type string');
         }
         this.field = field;
         this.config = config;
+        this.nestedTypes = nestedTypes;
     }
 
     abstract toESMapping(version?: number): TypeESMapping;
