+++ packages/data-types/test/data-type-es-spec.ts
@@ -1,0 +1,135 @@
+import 'jest-extended';
+import {
+    DataType, DataTypeConfig, LATEST_VERSION
+} from '../src';
+
+describe('DataType (elasticsearch)', () => {
+    describe('->toESMapping', () => {
+        it('elasticsearch mapping requires providing a type name', () => {
+            const typeConfig: DataTypeConfig = {
+                version: LATEST_VERSION,
+                fields: {
+                    hello: { type: 'Text' },
+                    location: { type: 'GeoPoint' },
+                    date: { type: 'Date' },
+                    ip: { type: 'IP' },
+                    someNum: { type: 'Long' },
+                },
+            };
+
+            try {
+                // @ts-ignore
+                new DataType(typeConfig).toESMapping();
+            } catch (err) {
+                expect(err.message).toInclude("Cannot destructure property `typeName` of 'undefined' or 'null'.");
+            }
+        });
+        it('can create an elasticsearch mapping', () => {
+            const typeConfig: DataTypeConfig = {
+                version: LATEST_VERSION,
+                fields: {
+                    hello: { type: 'Text' },
+                    location: { type: 'GeoPoint' },
+                    date: { type: 'Date' },
+                    ip: { type: 'IP' },
+                    someNum: { type: 'Long' },
+                },
+            };
+
+            const results = {
+                mappings: {
+                    events: {
+                        _all: {
+                            enabled: false,
+                        },
+                        dynamic: false,
+                        properties: {
+                            hello: { type: 'text' },
+                            location: { type: 'geo_point' },
+                            date: { type: 'date' },
+                            ip: { type: 'ip' },
+                            someNum: { type: 'long' },
+                        },
+                    },
+                },
+                settings: {},
+            };
+
+            const dataType = new DataType(typeConfig);
+
+            expect(dataType.toESMapping({ typeName: 'events' })).toEqual(results);
+        });
+
+        it('can add additional settings to a elasticsearch mapping', () => {
+            const typeConfig: DataTypeConfig = {
+                version: LATEST_VERSION,
+                fields: {
+                    hello: { type: 'KeywordTokensCaseInsensitive' },
+                    location: { type: 'GeoPoint' },
+                    date: { type: 'Date' },
+                    ip: { type: 'IP' },
+                    someNum: { type: 'Long' },
+                },
+            };
+
+            const overrides = {
+                settings: {
+                    'index.number_of_shards': 5,
+                    'index.number_of_replicas': 1,
+                    analysis: {
+                        analyzer: {
+                            some_other_analyzer: {
+                                tokenizer: 'keyword',
+                                filter: 'lowercase',
+                            },
+                        },
+                    },
+                },
+            };
+
+            const dataType = new DataType(typeConfig, 'events');
+            expect(dataType.toESMapping({ overrides })).toEqual({
+                mappings: {
+                    events: {
+                        _all: {
+                            enabled: false,
+                        },
+                        dynamic: false,
+                        properties: {
+                            hello: {
+                                type: 'text',
+                                analyzer: 'lowercase_keyword_analyzer',
+                                fields: {
+                                    tokens: {
+                                        analyzer: 'standard',
+                                        type: 'text',
+                                    },
+                                },
+                            },
+                            location: { type: 'geo_point' },
+                            date: { type: 'date' },
+                            ip: { type: 'ip' },
+                            someNum: { type: 'long' },
+                        },
+                    },
+                },
+                settings: {
+                    'index.number_of_shards': 5,
+                    'index.number_of_replicas': 1,
+                    analysis: {
+                        analyzer: {
+                            lowercase_keyword_analyzer: {
+                                tokenizer: 'keyword',
+                                filter: 'lowercase',
+                            },
+                            some_other_analyzer: {
+                                tokenizer: 'keyword',
+                                filter: 'lowercase',
+                            },
+                        },
+                    },
+                },
+            });
+        });
+    });
+});
